requirements.txt
absl-py==0.12.0
altgraph==0.17
astunparse==1.6.3
auto-py-to-exe==2.8.0
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Eel==0.12.4
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gevent==21.1.2
gevent-websocket==0.10.1
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.3
pefile==2019.4.18
Pillow==8.1.2
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
python-dateutil==2.8.1
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
